Add translations for admin login
