+++ src/Screens/Onboarding/Welcome.tsx
@@ -11,7 +11,8 @@ import useLayout from '../../lib/use-layout';
 import Card from '../components/Card';
 import SosBanner from '../components/CreatedBySosBanner';
 import AppTitle from '../components/AppTitle';
-import Message, { MessageWithLink } from '../components/Message';
+import Message from '../components/Message';
+import { MessageWithLink } from '../components/Link';
 import MessageButton from '../components/MessageButton';
 import fonts from '../components/fonts';
 import colors from '../components/colors';
