+++ src/Screens/components/Message.tsx
@@ -3,8 +3,6 @@ import RN from 'react-native';
 
 import * as localization from '../../localization';
 
-import Link from './Link';
-
 interface Props extends RN.TextProps {
   id: localization.MessageId;
 }
@@ -20,26 +18,6 @@ export const AnimatedMessage = ({
   style?: any;
 }) => (
   <RN.Animated.Text {...textProps}>{localization.trans(id)}</RN.Animated.Text>
-);
-
-interface MessageLinkProps extends RN.TextProps {
-  id: localization.MessageId;
-  link: {
-    linkName: localization.MessageId;
-    url: string;
-    style?: RN.StyleProp<RN.ViewStyle>;
-  };
-}
-
-export const MessageWithLink = ({
-  id,
-  link,
-  ...textProps
-}: MessageLinkProps) => (
-  <RN.Text {...textProps}>
-    {localization.trans(id)}
-    <Link {...link} />
-  </RN.Text>
 );
 
 export default Message;
