+++ src/localization/fi.ts
@@ -191,6 +191,9 @@ export const messages = {
   'onboarding.privacyPolicy.title': 'Yksityisyydensuoja',
 
   'onboarding.sign.in': 'Kirjaudu',
+  'onboarding.sign.in.admin': 'Admin-sisäänkirjautuminen',
+  'onboarding.sign.admin.info': 'Etkö ole admin?',
+  'onboarding.sign.admin.link': 'Kirjaudu sisään tavallisena käyttäjänä',
   'onboarding.sign.up': 'Rekisteröidy',
 
   'onboarding.signIn.button': 'Kirjaudu',
